A lot of stuff
Updates and additions galore

+++ Class Features/Relentless Rage.js
@@ -1,21 +1,20 @@
 async function wait(ms) { return new Promise(resolve => { setTimeout(resolve, ms); }); }
 const lastArg = args[args.length - 1];
 const actorM = canvas.tokens.get(lastArg.tokenId).actor;
-const DC = DAE.getFlag(actorM, "RelentlessRage");
-const hasEffectApplied = await game.dfreds.effectInterface.hasEffectApplied('Rage', actorM.uuid);
+const rageOn = await game.dfreds.effectInterface.hasEffectApplied('Rage', actorM.uuid);
 let effectData = lastArg.efData;
-console.log(args);
 
 if (args[0] == "off" && lastArg["expiry-reason"] == 'midi-qol:rest') {	
 	DAE.setFlag(actorM, "RelentlessRage", 10);
 	await actorM.createEmbeddedDocuments('ActiveEffect', [effectData]);
 }
 
-if (args[0] == "off" && lastArg["expiry-reason"] == 'midi-qol:zeroHP' && hasEffectApplied) {
+if (args[0] == "off" && lastArg["expiry-reason"] == 'midi-qol:zeroHP' && rageOn) {
 	await wait(1000); // needed cause sometimes updates too fast
 
 	const itemM = await actorM.items.getName("Relentless Rage");
 	const workflowItemData = duplicate(itemM);
+	const DC = DAE.getFlag(actorM, "RelentlessRage");
 	workflowItemData.system.save.dc = DC;
 
 	setProperty(workflowItemData, "flags.itemacro", {});
@@ -32,11 +31,11 @@ if (args[0] == "off" && lastArg["expiry-reason"] == 'midi-qol:zeroHP' && hasEffe
 		newDC = DC + 5;
 		DAE.setFlag(actorM, "RelentlessRage", newDC);
 		let prone = actorM.effects.find(e => e.label == "Prone");
-		prone.delete();
+		if (prone) prone.delete();
 	}
 	await actorM.createEmbeddedDocuments('ActiveEffect', [effectData]);
 }
 
-if (args[0] == "off" && lastArg["expiry-reason"] == 'midi-qol:zeroHP' && !hasEffectApplied) {
+if (args[0] == "off" && lastArg["expiry-reason"] == 'midi-qol:zeroHP' && !rageOn) {
 	await actorM.createEmbeddedDocuments('ActiveEffect', [effectData]);
 }
